+++ crates/core/src/errors.rs
@@ -12,24 +12,15 @@ use tokio::task::JoinError;
 
 use crate::hive::node::{Name, SwitchToConfigurationGoal};
 
-#[cfg(debug_assertions)]
-const DOCS_URL: &str = "http://localhost:5173/reference/errors.html";
-#[cfg(not(debug_assertions))]
-const DOCS_URL: &str = "https://wire.althaea.zone/reference/errors.html";
-
 #[derive(Debug, Diagnostic, Error)]
 pub enum KeyError {
-    #[diagnostic(
-        code(wire::key::File),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::key::File))]
     #[error("error reading file")]
     File(#[source] std::io::Error),
 
     #[diagnostic(
         code(wire::key::SpawningCommand),
-        help("Ensure wire has the correct $PATH for this command"),
-        url("{DOCS_URL}#{}", self.code().unwrap())
+        help("Ensure wire has the correct $PATH for this command")
     )]
     #[error("error spawning key command")]
     CommandSpawnError {
@@ -43,10 +34,7 @@ pub enum KeyError {
         command_span: Option<SourceSpan>,
     },
 
-    #[diagnostic(
-        code(wire::key::Resolving),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::key::Resolving))]
     #[error("Error resolving key command child process")]
     CommandResolveError {
         #[source]
@@ -56,24 +44,17 @@ pub enum KeyError {
         command: String,
     },
 
-    #[diagnostic(
-        code(wire::key::CommandExit),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::key::CommandExit))]
     #[error("key command failed with status {}: {}", .0,.1)]
     CommandError(ExitStatus, String),
 
-    #[diagnostic(
-        code(wire::key::Empty),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::key::Empty))]
     #[error("Command list empty")]
     Empty,
 
     #[diagnostic(
         code(wire::key::ParseKeyPermissions),
-        help("Refer to the documentation for the format of key file permissions."),
-        url("{DOCS_URL}#{}", self.code().unwrap())
+        help("Refer to the documentation for the format of key file permissions.")
     )]
     #[error("Failed to parse key permissions")]
     ParseKeyPermissions(#[source] ParseIntError),
@@ -81,10 +62,7 @@ pub enum KeyError {
 
 #[derive(Debug, Diagnostic, Error)]
 pub enum ActivationError {
-    #[diagnostic(
-        code(wire::activation::SwitchToConfiguration),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::activation::SwitchToConfiguration))]
     #[error("failed to run switch-to-configuration {0} on node {1}")]
     SwitchToConfigurationError(SwitchToConfigurationGoal, Name, #[source] CommandError),
 }
@@ -95,8 +73,7 @@ pub enum NetworkError {
         code(wire::network::HostUnreachable),
         help(
             "If you failed due to a fault in DNS, note that a node can have multiple targets defined."
-        ),
-        url("{DOCS_URL}#{}", self.code().unwrap())
+        )
     )]
     #[error("Cannot reach host {host}")]
     HostUnreachable {
@@ -105,17 +82,11 @@ pub enum NetworkError {
         source: CommandError,
     },
 
-    #[diagnostic(
-        code(wire::network::HostUnreachableAfterReboot),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::network::HostUnreachableAfterReboot))]
     #[error("Failed to get regain connection to {0} after activation.")]
     HostUnreachableAfterReboot(String),
 
-    #[diagnostic(
-        code(wire::network::HostsExhausted),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::network::HostsExhausted))]
     #[error("Ran out of contactable hosts")]
     HostsExhausted,
 }
@@ -126,32 +97,25 @@ pub enum HiveInitialisationError {
         code(wire::hive_init::NoHiveFound),
         help(
             "Double check the path is correct. You can adjust the hive path with `--path` when the hive lies outside of the CWD."
-        ),
-        url("{DOCS_URL}#{}", self.code().unwrap())
+        )
     )]
     #[error("No hive could be found in {}", .0.display())]
     NoHiveFound(PathBuf),
 
     #[diagnostic(
         code(wire::hive_init::Parse),
-        help("If you cannot resolve this problem, please create an issue."),
-        url("{DOCS_URL}#{}", self.code().unwrap())
+        help("If you cannot resolve this problem, please create an issue.")
     )]
     #[error("Failed to parse internal wire json.")]
     ParseEvaluateError(#[source] serde_json::Error),
 
-    #[diagnostic(
-        code(wire::hive_init::ParsePrefetch),
-        help("please create an issue."),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::hive_init::ParsePrefetch), help("please create an issue."))]
     #[error("Failed to parse `nix flake prefetch --json`.")]
     ParsePrefetchError(#[source] serde_json::Error),
 
     #[diagnostic(
         code(wire::hive_init::NodeDoesNotExist),
-        help("Please create an issue!"),
-        url("{DOCS_URL}#{}", self.code().unwrap())
+        help("Please create an issue!")
     )]
     #[error("node {0} not exist in hive")]
     NodeDoesNotExist(String),
@@ -159,109 +123,69 @@ pub enum HiveInitialisationError {
 
 #[derive(Debug, Diagnostic, Error)]
 pub enum HiveLocationError {
-    #[diagnostic(
-        code(wire::hive_location::MalformedPath),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::hive_location::MalformedPath))]
     #[error("Path was malformed: {}", .0.display())]
     MalformedPath(PathBuf),
 
-    #[diagnostic(
-        code(wire::hive_location::Malformed),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::hive_location::Malformed))]
     #[error("--path was malformed")]
     Malformed(#[source] FlakeRefError),
 
-    #[diagnostic(
-        code(wire::hive_location::TypeUnsupported),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::hive_location::TypeUnsupported))]
     #[error("The flakref had an unsupported type: {:#?}", .0)]
     TypeUnsupported(Box<FlakeRef>),
 }
 
 #[derive(Debug, Diagnostic, Error)]
 pub enum CommandError {
-    #[diagnostic(
-        code(wire::command::TermAttrs),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::command::TermAttrs))]
     #[error("Failed to set PTY attrs")]
     TermAttrs(#[source] nix::errno::Errno),
 
-    #[diagnostic(
-        code(wire::command::PosixPipe),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::command::PosixPipe))]
     #[error("There was an error in regards to a pipe")]
     PosixPipe(#[source] nix::errno::Errno),
 
     /// Error wrapped around `portable_pty`'s anyhow
     /// errors
-    #[diagnostic(
-        code(wire::command::PortablePty),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::command::PortablePty))]
     #[error("There was an error from the portable_pty crate")]
     PortablePty(#[source] anyhow::Error),
 
-    #[diagnostic(
-        code(wire::command::Joining),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::command::Joining))]
     #[error("Failed to join on some tokio task")]
     JoinError(#[source] JoinError),
 
-    #[diagnostic(
-        code(wire::command::WaitForStatus),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::command::WaitForStatus))]
     #[error("Failed to wait for the child's status")]
     WaitForStatus(#[source] std::io::Error),
 
     #[diagnostic(
         code(wire::detached::NoHandle),
-        help("This should never happen, please create an issue!"),
-        url("{DOCS_URL}#{}", self.code().unwrap())
+        help("This should never happen, please create an issue!")
     )]
     #[error("There was no handle to child io")]
     NoHandle,
 
-    #[diagnostic(
-        code(wire::command::WritingClientStdout),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::command::WritingClientStdout))]
     #[error("Failed to write to client stderr.")]
     WritingClientStderr(#[source] std::io::Error),
 
-    #[diagnostic(
-        code(wire::command::WritingMasterStdin),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::command::WritingMasterStdin))]
     #[error("Failed to write to PTY master stdout.")]
     WritingMasterStdout(#[source] std::io::Error),
 
-    #[diagnostic(
-        code(wire::command::Recv),
-        url("{DOCS_URL}#{}", self.code().unwrap()),
-        help("please create an issue!"),
-    )]
+    #[diagnostic(code(wire::command::Recv), help("please create an issue!"))]
     #[error("Failed to receive a message from the begin channel")]
     RecvError(#[source] RecvError),
 
-    #[diagnostic(
-        code(wire::command::ThreadPanic),
-        url("{DOCS_URL}#{}", self.code().unwrap()),
-        help("please create an issue!"),
-    )]
+    #[diagnostic(code(wire::command::ThreadPanic), help("please create an issue!"))]
     #[error("Thread panicked")]
     ThreadPanic,
 
     #[diagnostic(
         code(wire::command::CommandFailed),
-        url("{DOCS_URL}#{}", self.code().unwrap()),
-        help("`nix` commands are filtered, run with -vvv to view all"),
+        help("`nix` commands are filtered, run with -vvv to view all")
     )]
     #[error("{command_ran} failed ({reason}) with {code} (last 20 lines):\n{logs}")]
     CommandFailed {
@@ -271,24 +195,15 @@ pub enum CommandError {
         reason: &'static str,
     },
 
-    #[diagnostic(
-        code(wire::command::RuntimeDirectory),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::command::RuntimeDirectory))]
     #[error("error creating $XDG_RUNTIME_DIR/wire")]
     RuntimeDirectory(#[source] std::io::Error),
 
-    #[diagnostic(
-        code(wire::command::RuntimeDirectoryMissing),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::command::RuntimeDirectoryMissing))]
     #[error("$XDG_RUNTIME_DIR could not be used.")]
     RuntimeDirectoryMissing(#[source] std::env::VarError),
 
-    #[diagnostic(
-        code(wire::command::OneshotRecvError),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::command::OneshotRecvError))]
     #[error("Error waiting for begin message")]
     OneshotRecvError(#[source] tokio::sync::oneshot::error::RecvError),
 }
@@ -323,10 +238,7 @@ pub enum HiveLibError {
         KeyError,
     ),
 
-    #[diagnostic(
-        code(wire::BuildNode),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::BuildNode))]
     #[error("failed to build node {name}")]
     NixBuildError {
         name: Name,
@@ -334,10 +246,7 @@ pub enum HiveLibError {
         source: CommandError,
     },
 
-    #[diagnostic(
-        code(wire::CopyPath),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::CopyPath))]
     #[error("failed to copy path {path} to node {name}")]
     NixCopyError {
         name: Name,
@@ -360,17 +269,11 @@ pub enum HiveLibError {
         help: Option<Box<String>>,
     },
 
-    #[diagnostic(
-        code(wire::Encoding),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::Encoding))]
     #[error("error encoding length delimited data")]
     Encoding(#[source] std::io::Error),
 
-    #[diagnostic(
-        code(wire::SIGINT),
-        url("{DOCS_URL}#{}", self.code().unwrap())
-    )]
+    #[diagnostic(code(wire::SIGINT))]
     #[error("SIGINT received, shut down")]
     Sigint,
 }